Move FeatureFlags to Ultima.Data namespace
Because data does not belong in the Network namespace!
