+++ dev/Ultima/Data/FeatureFlags.cs
@@ -1,6 +1,17 @@
-﻿using System;
+﻿/***************************************************************************
+ *   FeatureFlags.cs
+ *   Copyright (c) 2015 UltimaXNA Development Team
+ * 
+ *   This program is free software; you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation; either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ ***************************************************************************/
 
-namespace UltimaXNA.Ultima.Network.Server
+using System;
+
+namespace UltimaXNA.Ultima.Data
 {
     [Flags]
     public enum FeatureFlags : uint
